clean up code

+++ BufferExcelWriter/WorkBookDfn.cs
@@ -1,18 +1,28 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
-using System.Transactions;
 using System.Xml;
 
 namespace BufferExcelWriter
 {
     public class WorkBookDfn : IDisposable
     {
+        internal const String WorksheetDefaultHeaders =
+            "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?> <worksheet xmlns=\"http://schemas.openxmlformats.org/spreadsheetml/2006/main\" xmlns:r=\"http://schemas.openxmlformats.org/officeDocument/2006/relationships\" xmlns:mc=\"http://schemas.openxmlformats.org/markup-compatibility/2006\" mc:Ignorable=\"x14ac xr xr2 xr3\" xmlns:x14ac=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac\" xmlns:xr=\"http://schemas.microsoft.com/office/spreadsheetml/2014/revision\" xmlns:xr2=\"http://schemas.microsoft.com/office/spreadsheetml/2015/revision2\" xmlns:xr3=\"http://schemas.microsoft.com/office/spreadsheetml/2016/revision3\" xr:uid=\"{00000000-0001-0000-0000-000000000000}\"> <sheetData>";
+
+        internal const String WorksheetDefaultFooter = "</worksheet>";
+        internal const String SheetDataDefaultFooter = "</sheetData>";
+
+        private readonly Dictionary<Int32, Int32> _rowOffsetDic = new Dictionary<Int32, Int32>();
+        internal String OutPutFilePath;
+        internal Stream OutputStream;
+        internal String WorkingFolder;
+
         public WorkBookDfn()
         {
             WorkingFolder = Guid.NewGuid().ToString("N");
@@ -21,56 +31,52 @@ namespace BufferExcelWriter
                 var existDir = new DirectoryInfo(WorkingFolder);
                 existDir.Delete(true);
             }
-            Directory.CreateDirectory(WorkingFolder);
-            using (var fs = File.OpenRead(@"C:\Users\itok\Desktop\advTemplate.zip"))
 
+            Directory.CreateDirectory(WorkingFolder);
+            OutPutFilePath = WorkingFolder + ".zip";
+            if (File.Exists(OutPutFilePath)) File.Delete(OutPutFilePath);
+            var assembly = Assembly.GetExecutingAssembly();
+            using (var fs = assembly.GetManifestResourceStream("BufferExcelWriter.exceltemplate"))
             {
+                if (fs == null) throw new FileNotFoundException("BufferExcelWriter.exceltemplate");
                 var zipFile = new ZipArchive(fs);
                 zipFile.ExtractToDirectory(WorkingFolder);
             }
 
             Sheets = new List<WorkSheetDfn>();
             FolderEntry = new FolderEntry(WorkingFolder);
-            Stopwatch = new Stopwatch();
         }
 
-        private readonly Dictionary<Int32, Int32> _rowOffsetDic = new Dictionary<Int32, Int32>();
-        internal String WorkingFolder;
-        public Stopwatch Stopwatch { get; set; }
         internal FolderEntry FolderEntry { get; set; }
         public IList<WorkSheetDfn> Sheets { get; set; }
 
-        public void CleanSheetsBuffer()
+        /// <summary>
+        ///clean temp folder and file
+        /// </summary>
+        public void Dispose()
         {
-            foreach (var sheet in Sheets)
-            {
-                sheet.BufferedRows.Clear();
-                GC.Collect();
-            }
+            foreach (var sheet in Sheets) sheet.StreamWriter.Dispose();
+            if (File.Exists(OutPutFilePath)) File.Delete(OutPutFilePath);
+            OutputStream.Dispose();
         }
 
         public async Task OpenWriteExcelAsync()
         {
-            for (int i = 0; i < Sheets.Count; i++)
+            for (var i = 0; i < Sheets.Count; i++)
             {
                 var currentSheet = Sheets[i];
-                if (String.IsNullOrWhiteSpace(currentSheet.Name))
-                {
-                    currentSheet.Name = $"Sheet{i + 1}";
-                }
+                if (String.IsNullOrWhiteSpace(currentSheet.Name)) currentSheet.Name = $"Sheet{i + 1}";
                 currentSheet.SheetNum = i + 1;
             }
 
             foreach (var sheet in Sheets)
             {
                 _rowOffsetDic[sheet.SheetNum] = 1;
+
                 #region Update [Content_Types].xml
 
                 var contentTypeEntry = FolderEntry.GetEntry("[Content_Types].xml");
-                if (contentTypeEntry == null)
-                {
-                    throw new FileNotFoundException("[Content_Types].xml");
-                }
+                if (contentTypeEntry == null) throw new FileNotFoundException("[Content_Types].xml");
 
                 using (var contentTypeStream = contentTypeEntry.Open())
                 {
@@ -89,6 +95,7 @@ namespace BufferExcelWriter
                                 "application/vnd.openxmlformats-officedocument.spreadsheetml.worksheet+xml");
                             types.AppendChild(element);
                         }
+
                         contentTypeStream.Position = 0;
                         contentTypeStream.SetLength(0);
                         doc.Save(contentTypeStream);
@@ -100,12 +107,9 @@ namespace BufferExcelWriter
 
                 #region Update xl/_rels/workbook.xml.rels
 
-                string identifier = "rId";
+                var identifier = "rId";
                 var relsEntry = FolderEntry.GetEntry("xl/_rels/workbook.xml.rels");
-                if (relsEntry == null)
-                {
-                    throw new FileNotFoundException("xl/_rels/workbook.xml.rels");
-                }
+                if (relsEntry == null) throw new FileNotFoundException("xl/_rels/workbook.xml.rels");
 
                 using (var relsStream = relsEntry.Open())
                 {
@@ -134,18 +138,13 @@ namespace BufferExcelWriter
                     }
                 }
 
-
-
                 #endregion
 
 
                 #region Update xl/workbook.xml
 
                 var workbookEntry = FolderEntry.GetEntry("xl/workbook.xml");
-                if (workbookEntry == null)
-                {
-                    throw new FileNotFoundException("xl/workbook.xml");
-                }
+                if (workbookEntry == null) throw new FileNotFoundException("xl/workbook.xml");
 
                 using (var workbookStream = workbookEntry.Open())
                 {
@@ -180,10 +179,7 @@ namespace BufferExcelWriter
                 #region Update docProps/app.xml
 
                 var appEntry = FolderEntry.GetEntry("docProps/app.xml");
-                if (appEntry == null)
-                {
-                    throw new FileNotFoundException("docProps/app.xml Not Found in FolderEntry");
-                }
+                if (appEntry == null) throw new FileNotFoundException("docProps/app.xml Not Found in FolderEntry");
 
                 using (var appStream = appEntry.Open())
                 {
@@ -226,111 +222,83 @@ namespace BufferExcelWriter
                 sheet.StreamWriter = new StreamWriter(sheet.FileStream);
                 //sheet.FileStream.Position = sheet.FileStream.Length;
                 await sheet.StreamWriter.WriteAsync(WorksheetDefaultHeaders);
-                if (sheet.Header == null)
-                {
-                    throw new Exception("Header mustn't be null!");
-                }
+                if (sheet.Header == null) throw new Exception("Header mustn't be null!");
                 await sheet.StreamWriter.WriteAsync(sheet.Header.ToXmlString(1, sheet.Header, sheet.NullValStr));
                 _rowOffsetDic[sheet.SheetNum]++;
-
-
 
                 #endregion
             }
         }
 
-        public async Task FlushBufferRowsAsync()
+        /// <summary>
+        ///     Flush bufferedRow data to disk and clean buffered row
+        /// </summary>
+        /// <returns></returns>
+        public async Task FlushBufferedRowsAsync(Boolean needGc = false)
         {
-            Stopwatch.Start();
             foreach (var sheet in Sheets)
             {
                 var sheetEntry = FolderEntry.GetEntry(sheet.GetEntryName());
-                if (sheetEntry == null)
-                {
-                    throw new Exception("Init needed!!");
-                }
+                if (sheetEntry == null) throw new Exception("Init needed!!");
 
                 var sb = new StringBuilder();
                 for (var rowIndex = 0; rowIndex < sheet.BufferedRows.Count(); rowIndex++)
                 {
                     var currentRow = sheet.BufferedRows[rowIndex];
-                    //var buffer = Encoding.UTF8.GetBytes(currentRow.ToXmlString(_rowOffsetDic[sheet.SheetNum],
-                    //    sheet.Header, sheet.NullValStr));
-                    //await sheetStream.WriteAsync(buffer, 0, buffer.Length);
                     sb.Append(currentRow.ToXmlString(_rowOffsetDic[sheet.SheetNum], sheet.Header,
                         sheet.NullValStr));
                     _rowOffsetDic[sheet.SheetNum]++;
                 }
+
                 await sheet.StreamWriter.WriteAsync(sb.ToString());
-
-                //await sheetStream.FlushAsync();
                 await sheet.StreamWriter.FlushAsync();
-
+                sheet.BufferedRows.Clear();
             }
-            Stopwatch.Stop();
 
-            Console.WriteLine(Stopwatch.Elapsed);
-            Stopwatch.Reset();
-
+            if (needGc)
+            {
+                GC.Collect();
+            }
         }
 
+        /// <summary>
+        ///     write end info to file and get file stream back
+        /// </summary>
+        /// <returns></returns>
         public async Task<Stream> CloseExcelAndGetStreamAsync()
         {
-
             foreach (var sheet in Sheets)
             {
                 var sheetEntry = FolderEntry.GetEntry(sheet.GetEntryName());
-                if (sheetEntry == null)
-                {
-                    throw new FileNotFoundException(sheet.GetEntryName());
-                }
+                if (sheetEntry == null) throw new FileNotFoundException(sheet.GetEntryName());
                 await sheet.StreamWriter.WriteAsync(SheetDataDefaultFooter);
                 await sheet.StreamWriter.WriteAsync(WorksheetDefaultFooter);
                 await sheet.StreamWriter.FlushAsync();
                 sheet.StreamWriter.Dispose();
             }
-            ZipFile.CreateFromDirectory(WorkingFolder, WorkingFolder + ".zip", CompressionLevel.Fastest, true);
-            return File.Open(WorkingFolder + ".zip", FileMode.Open);
-        }
 
-
-
-        public const String WorksheetDefaultHeaders = "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?> <worksheet xmlns=\"http://schemas.openxmlformats.org/spreadsheetml/2006/main\" xmlns:r=\"http://schemas.openxmlformats.org/officeDocument/2006/relationships\" xmlns:mc=\"http://schemas.openxmlformats.org/markup-compatibility/2006\" mc:Ignorable=\"x14ac xr xr2 xr3\" xmlns:x14ac=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac\" xmlns:xr=\"http://schemas.microsoft.com/office/spreadsheetml/2014/revision\" xmlns:xr2=\"http://schemas.microsoft.com/office/spreadsheetml/2015/revision2\" xmlns:xr3=\"http://schemas.microsoft.com/office/spreadsheetml/2016/revision3\" xr:uid=\"{00000000-0001-0000-0000-000000000000}\"> <sheetData>";
-        public const String WorksheetDefaultFooter = "</worksheet>";
-        public const String SheetDataDefaultFooter = "</sheetData>";
-
-        public void Dispose()
-        {
-            foreach (var sheet in Sheets)
-            {
-                sheet.StreamWriter.Dispose();
-            }
-            if (Directory.Exists(WorkingFolder))
-            {
-                Directory.Delete(WorkingFolder, true);
-            }
-            if (File.Exists(WorkingFolder + ".zip"))
-            {
-
-                File.Delete(WorkingFolder + ".zip");
-            }
+            ZipFile.CreateFromDirectory(WorkingFolder, OutPutFilePath, CompressionLevel.Optimal, false);
+            if (Directory.Exists(WorkingFolder)) Directory.Delete(WorkingFolder, true);
+            OutputStream = File.Open(OutPutFilePath, FileMode.Open);
+            return OutputStream;
         }
     }
+
     internal class FolderEntry
     {
         private readonly String _dirPath;
 
-        public FolderEntry(String dirPath)
+        internal FolderEntry(String dirPath)
         {
             _dirPath = dirPath;
         }
 
-        public FileEntry GetEntry(String filename)
+        internal FileEntry GetEntry(String filename)
         {
             return new FileEntry(new FileInfo(Path.Combine(_dirPath, filename)));
         }
 
-        public FileEntry CreateEntry(String filename)
+        internal FileEntry CreateEntry(String filename)
         {
             return GetEntry(filename);
         }
@@ -339,12 +307,13 @@ namespace BufferExcelWriter
     internal class FileEntry
     {
         private readonly FileInfo _fileInfo;
-        public FileEntry(FileInfo fileInfo)
+
+        internal FileEntry(FileInfo fileInfo)
         {
             _fileInfo = fileInfo;
         }
 
-        public Stream Open()
+        internal Stream Open()
         {
             return _fileInfo.Open(FileMode.OpenOrCreate);
         }
